Move ToolUiLinkImpl to own file
